pyproject.toml
[tool.poetry]
name = "schoolmaterials"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
mysql-connector = "^2.2.9"
python-dotenv = "^0.15.0"
flask = "^1.1.2"
discord-webhook = "^0.12.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
